Create LoggerInjector and move the logic there
