+++ src/Bridge/Monolog/tests/FactoryTest.php
@@ -17,9 +17,9 @@ use Spiral\Config\ConfigManager;
 use Spiral\Config\ConfiguratorInterface;
 use Spiral\Config\LoaderInterface;
 use Spiral\Core\Container;
+use Spiral\Logger\Attribute\LoggerChannel;
 use Spiral\Logger\ListenerRegistry;
 use Spiral\Logger\LogsInterface;
-use Spiral\Monolog\Attribute\LoggerChannel;
 use Spiral\Monolog\Bootloader\MonologBootloader;
 use Spiral\Monolog\Config\MonologConfig;
 use Spiral\Monolog\LogFactory;
